# optimize etcd3 getlist