<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?

<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->

#### What this PR does / why we need it:

Faster etcd3 get-list.

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

#### Special notes for your reviewer:

Here are my own benchmark results. FIY...

```bash
$ cat etcd.before                                                                                                   
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/pkg/storage/etcd3
cpu: AMD EPYC 7K62 48-Core Processor
BenchmarkStore_GetList/list_10_objects-16         	    3861	    320458 ns/op	   71473 B/op	     637 allocs/op
BenchmarkStore_GetList/list_10_objects-16         	    3878	    318080 ns/op	   71348 B/op	     637 allocs/op
BenchmarkStore_GetList/list_100_objects-16        	     957	   1459241 ns/op	  581732 B/op	    3535 allocs/op
BenchmarkStore_GetList/list_100_objects-16        	     976	   1288252 ns/op	  576384 B/op	    3534 allocs/op
BenchmarkStore_GetList/list_1000_objects-16       	     100	  12729213 ns/op	 6047508 B/op	   34832 allocs/op
BenchmarkStore_GetList/list_1000_objects-16       	     100	  11991752 ns/op	 6050610 B/op	   34939 allocs/op
BenchmarkStore_GetList/list_10000_objects-16      	      10	 115570547 ns/op	60728416 B/op	  359802 allocs/op
BenchmarkStore_GetList/list_10000_objects-16      	      12	 111636096 ns/op	55874789 B/op	  359855 allocs/op
BenchmarkStore_GetList/list_50000_objects-16                  2	         513074251 ns/op	264123340 B/op	 1801004 allocs/op
BenchmarkStore_GetList/list_50000_objects-16      	      2	         528685710 ns/op	264113632 B/op	 1800347 allocs/op
PASS
ok  	k8s.io/apiserver/pkg/storage/etcd3	85.554s

$ cat etcd.after
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/pkg/storage/etcd3
cpu: AMD EPYC 7K62 48-Core Processor
BenchmarkStore_GetList/list_10_objects-16         	    3796	    298688 ns/op	   54006 B/op	     643 allocs/op
BenchmarkStore_GetList/list_10_objects-16         	    3746	    295224 ns/op	   53952 B/op	     643 allocs/op
BenchmarkStore_GetList/list_100_objects-16        	    1050	   1147579 ns/op	  377699 B/op	    3458 allocs/op
BenchmarkStore_GetList/list_100_objects-16        	    1038	   1124952 ns/op	  376834 B/op	    3446 allocs/op
BenchmarkStore_GetList/list_1000_objects-16       	     123	   9566107 ns/op	 3603900 B/op	   33881 allocs/op
BenchmarkStore_GetList/list_1000_objects-16       	     126	   9449522 ns/op	 3606441 B/op	   33941 allocs/op
BenchmarkStore_GetList/list_10000_objects-16      	      12	  94111233 ns/op	37636011 B/op	  349829 allocs/op
BenchmarkStore_GetList/list_10000_objects-16      	      12	  93095770 ns/op	37818962 B/op	  350481 allocs/op
BenchmarkStore_GetList/list_50000_objects-16      	       3	 466840470 ns/op	189827453 B/op	 1749952 allocs/op
BenchmarkStore_GetList/list_50000_objects-16      	       3	 465690673 ns/op	190029832 B/op	 1750305 allocs/op
PASS
ok  	k8s.io/apiserver/pkg/storage/etcd3	114.789s

$ benchcmp etcd.before etcd.after                                                                                                    
benchmark                                        old ns/op     new ns/op     delta
BenchmarkStore_GetList/list_10_objects-16        320458        298688        -6.79%
BenchmarkStore_GetList/list_10_objects-16        318080        295224        -7.19%
BenchmarkStore_GetList/list_100_objects-16       1459241       1147579       -21.36%
BenchmarkStore_GetList/list_100_objects-16       1288252       1124952       -12.68%
BenchmarkStore_GetList/list_1000_objects-16      12729213      9566107       -24.85%
BenchmarkStore_GetList/list_1000_objects-16      11991752      9449522       -21.20%
BenchmarkStore_GetList/list_10000_objects-16     115570547     94111233      -18.57%
BenchmarkStore_GetList/list_10000_objects-16     111636096     93095770      -16.61%
BenchmarkStore_GetList/list_50000_objects-16     513074251     466840470     -9.01%
BenchmarkStore_GetList/list_50000_objects-16     528685710     465690673     -11.92%

benchmark                                        old allocs     new allocs     delta
BenchmarkStore_GetList/list_10_objects-16        637            643            +0.94%
BenchmarkStore_GetList/list_10_objects-16        637            643            +0.94%
BenchmarkStore_GetList/list_100_objects-16       3535           3458           -2.18%
BenchmarkStore_GetList/list_100_objects-16       3534           3446           -2.49%
BenchmarkStore_GetList/list_1000_objects-16      34832          33881          -2.73%
BenchmarkStore_GetList/list_1000_objects-16      34939          33941          -2.86%
BenchmarkStore_GetList/list_10000_objects-16     359802         349829         -2.77%
BenchmarkStore_GetList/list_10000_objects-16     359855         350481         -2.60%
BenchmarkStore_GetList/list_50000_objects-16     1801004        1749952        -2.83%
BenchmarkStore_GetList/list_50000_objects-16     1800347        1750305        -2.78%

benchmark                                        old bytes     new bytes     delta
BenchmarkStore_GetList/list_10_objects-16        71473         54006         -24.44%
BenchmarkStore_GetList/list_10_objects-16        71348         53952         -24.38%
BenchmarkStore_GetList/list_100_objects-16       581732        377699        -35.07%
BenchmarkStore_GetList/list_100_objects-16       576384        376834        -34.62%
BenchmarkStore_GetList/list_1000_objects-16      6047508       3603900       -40.41%
BenchmarkStore_GetList/list_1000_objects-16      6050610       3606441       -40.40%
BenchmarkStore_GetList/list_10000_objects-16     60728416      37636011      -38.03%
BenchmarkStore_GetList/list_10000_objects-16     55874789      37818962      -32.31%
BenchmarkStore_GetList/list_50000_objects-16     264123340     189827453     -28.13%
BenchmarkStore_GetList/list_50000_objects-16     264113632     190029832     -28.05%
```

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note

```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```