@sxllwx 2023-03-09
Update: 

````bash
# benchcmp 
benchmark                                            old ns/op      new ns/op      delta
BenchmarkStore_GetList/list_10_corev1.Pod-16         475431         453929         -4.52%
BenchmarkStore_GetList/list_10_corev1.Pod-16         464205         448608         -3.36%
BenchmarkStore_GetList/list_100_corev1.Pod-16        2815236        2575968        -8.50%
BenchmarkStore_GetList/list_100_corev1.Pod-16        2913277        2569991        -11.78%
BenchmarkStore_GetList/list_1000_corev1.Pod-16       26287639       24295101       -7.58%
BenchmarkStore_GetList/list_1000_corev1.Pod-16       26641453       23947811       -10.11%
BenchmarkStore_GetList/list_10000_corev1.Pod-16      252053668      238738248      -5.28%
BenchmarkStore_GetList/list_10000_corev1.Pod-16      272613894      234570819      -13.95%
BenchmarkStore_GetList/list_50000_corev1.Pod-16      1213607441     1214340396     +0.06%
BenchmarkStore_GetList/list_50000_corev1.Pod-16      1194805913     1197901030     +0.26%
BenchmarkStore_GetList/list_10_corev1.Node-16        477939         454682         -4.87%
BenchmarkStore_GetList/list_10_corev1.Node-16        484702         453219         -6.50%
BenchmarkStore_GetList/list_100_corev1.Node-16       2767087        2656915        -3.98%
BenchmarkStore_GetList/list_100_corev1.Node-16       2770697        2641017        -4.68%
BenchmarkStore_GetList/list_1000_corev1.Node-16      26196409       24679840       -5.79%
BenchmarkStore_GetList/list_1000_corev1.Node-16      26459178       24623285       -6.94%
BenchmarkStore_GetList/list_10000_corev1.Node-16     261074212      242936736      -6.95%
BenchmarkStore_GetList/list_10000_corev1.Node-16     254321802      243070395      -4.42%
BenchmarkStore_GetList/list_50000_corev1.Node-16     1231968582     1215763255     -1.32%
BenchmarkStore_GetList/list_50000_corev1.Node-16     1209250275     1207361772     -0.16%

benchmark                                            old allocs     new allocs     delta
BenchmarkStore_GetList/list_10_corev1.Pod-16         900            898            -0.22%
BenchmarkStore_GetList/list_10_corev1.Pod-16         887            902            +1.69%
BenchmarkStore_GetList/list_100_corev1.Pod-16        6119           5953           -2.71%
BenchmarkStore_GetList/list_100_corev1.Pod-16        6040           6033           -0.12%
BenchmarkStore_GetList/list_1000_corev1.Pod-16       60508          59499          -1.67%
BenchmarkStore_GetList/list_1000_corev1.Pod-16       60529          59523          -1.66%
BenchmarkStore_GetList/list_10000_corev1.Pod-16      609795         600322         -1.55%
BenchmarkStore_GetList/list_10000_corev1.Pod-16      610197         600522         -1.59%
BenchmarkStore_GetList/list_50000_corev1.Pod-16      3052665        3002197        -1.65%
BenchmarkStore_GetList/list_50000_corev1.Pod-16      3052297        3002216        -1.64%
BenchmarkStore_GetList/list_10_corev1.Node-16        955            934            -2.20%
BenchmarkStore_GetList/list_10_corev1.Node-16        937            934            -0.32%
BenchmarkStore_GetList/list_100_corev1.Node-16       6541           6455           -1.31%
BenchmarkStore_GetList/list_100_corev1.Node-16       6541           6453           -1.35%
BenchmarkStore_GetList/list_1000_corev1.Node-16      65397          64456          -1.44%
BenchmarkStore_GetList/list_1000_corev1.Node-16      65748          64562          -1.80%
BenchmarkStore_GetList/list_10000_corev1.Node-16     660805         649795         -1.67%
BenchmarkStore_GetList/list_10000_corev1.Node-16     659782         649671         -1.53%
BenchmarkStore_GetList/list_50000_corev1.Node-16     3302058        3252525        -1.50%
BenchmarkStore_GetList/list_50000_corev1.Node-16     3302582        3251713        -1.54%

benchmark                                            old bytes     new bytes     delta
BenchmarkStore_GetList/list_10_corev1.Pod-16         143617        105911        -26.25%
BenchmarkStore_GetList/list_10_corev1.Pod-16         143836        105992        -26.31%
BenchmarkStore_GetList/list_100_corev1.Pod-16        1271534       877053        -31.02%
BenchmarkStore_GetList/list_100_corev1.Pod-16        1382780       879686        -36.38%
BenchmarkStore_GetList/list_1000_corev1.Pod-16       12796879      8759114       -31.55%
BenchmarkStore_GetList/list_1000_corev1.Pod-16       12792836      8747185       -31.62%
BenchmarkStore_GetList/list_10000_corev1.Pod-16      118933402     88056604      -25.96%
BenchmarkStore_GetList/list_10000_corev1.Pod-16      118966072     88100177      -25.95%
BenchmarkStore_GetList/list_50000_corev1.Pod-16      441080400     444088304     +0.68%
BenchmarkStore_GetList/list_50000_corev1.Pod-16      441022056     444101616     +0.70%
BenchmarkStore_GetList/list_10_corev1.Node-16        109802        85715         -21.94%
BenchmarkStore_GetList/list_10_corev1.Node-16        117276        85604         -27.01%
BenchmarkStore_GetList/list_100_corev1.Node-16       927463        676136        -27.10%
BenchmarkStore_GetList/list_100_corev1.Node-16       931019        674296        -27.57%
BenchmarkStore_GetList/list_1000_corev1.Node-16      9474277       6661856       -29.68%
BenchmarkStore_GetList/list_1000_corev1.Node-16      9495070       6647492       -29.99%
BenchmarkStore_GetList/list_10000_corev1.Node-16     87340370      67068329      -23.21%
BenchmarkStore_GetList/list_10000_corev1.Node-16     87347420      67049544      -23.24%
BenchmarkStore_GetList/list_50000_corev1.Node-16     333308272     336621728     +0.99%
BenchmarkStore_GetList/list_50000_corev1.Node-16     333246248     336292280     +0.91%
````