@sxllwx 2023-03-13
```bash
benchmark                                                        old ns/op      new ns/op      delta
BenchmarkStore_GetList/pick_1000_pods_out_of_100000_pod-16       6269308270     6250230597     -0.30%
BenchmarkStore_GetList/pick_1000_pods_out_of_100000_pod-16       6287190432     6266327522     -0.33%
BenchmarkStore_GetList/pick_10000_Pods_out_of_100000_pod-16      6299670296     6303797752     +0.07%
BenchmarkStore_GetList/pick_10000_Pods_out_of_100000_pod-16      6317763922     6305945383     -0.19%
BenchmarkStore_GetList/pick_25000_Pods_out_of_100000_pod-16      6312331931     6352382247     +0.63%
BenchmarkStore_GetList/pick_25000_Pods_out_of_100000_pod-16      6352589171     6387865500     +0.56%
BenchmarkStore_GetList/pick_50000_Pods_out_of_100000_pod-16      6387484883     6376653645     -0.17%
BenchmarkStore_GetList/pick_50000_Pods_out_of_100000_pod-16      6426496866     6362955333     -0.99%
BenchmarkStore_GetList/pick_100000_Pods_out_of_100000_pod-16     6556018553     6428002054     -1.95%
BenchmarkStore_GetList/pick_100000_Pods_out_of_100000_pod-16     6544365720     6437920403     -1.63%

benchmark                                                        old allocs     new allocs     delta
BenchmarkStore_GetList/pick_1000_pods_out_of_100000_pod-16       5124489        5124203        -0.01%
BenchmarkStore_GetList/pick_1000_pods_out_of_100000_pod-16       5122806        5118873        -0.08%
BenchmarkStore_GetList/pick_10000_Pods_out_of_100000_pod-16      5202060        5200043        -0.04%
BenchmarkStore_GetList/pick_10000_Pods_out_of_100000_pod-16      5202554        5192521        -0.19%
BenchmarkStore_GetList/pick_25000_Pods_out_of_100000_pod-16      5338729        5310609        -0.53%
BenchmarkStore_GetList/pick_25000_Pods_out_of_100000_pod-16      5336929        5314654        -0.42%
BenchmarkStore_GetList/pick_50000_Pods_out_of_100000_pod-16      5564591        5513459        -0.92%
BenchmarkStore_GetList/pick_50000_Pods_out_of_100000_pod-16      5562133        5512893        -0.89%
BenchmarkStore_GetList/pick_100000_Pods_out_of_100000_pod-16     6015314        5912202        -1.71%
BenchmarkStore_GetList/pick_100000_Pods_out_of_100000_pod-16     6011989        5912739        -1.65%

benchmark                                                        old bytes      new bytes      delta
BenchmarkStore_GetList/pick_1000_pods_out_of_100000_pod-16       1518817664     1518214144     -0.04%
BenchmarkStore_GetList/pick_1000_pods_out_of_100000_pod-16       1518802912     1517903264     -0.06%
BenchmarkStore_GetList/pick_10000_Pods_out_of_100000_pod-16      1541939920     1528406264     -0.88%
BenchmarkStore_GetList/pick_10000_Pods_out_of_100000_pod-16      1542009352     1528227720     -0.89%
BenchmarkStore_GetList/pick_25000_Pods_out_of_100000_pod-16      1588886120     1545267600     -2.75%
BenchmarkStore_GetList/pick_25000_Pods_out_of_100000_pod-16      1588706968     1545447152     -2.72%
BenchmarkStore_GetList/pick_50000_Pods_out_of_100000_pod-16      1666975368     1573613744     -5.60%
BenchmarkStore_GetList/pick_50000_Pods_out_of_100000_pod-16      1666638248     1573729472     -5.57%
BenchmarkStore_GetList/pick_100000_Pods_out_of_100000_pod-16     1885756232     1630320928     -13.55%
BenchmarkStore_GetList/pick_100000_Pods_out_of_100000_pod-16     1885888072     1630109920     -13.56%
```